+++ src/renderer/components/Note.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Dropdown, Menu, Divider, Checkbox, Tooltip, Collapse, Typography, Button } from 'antd';
-import Icon, { UnorderedListOutlined, PlusOutlined, DeleteFilled, EllipsisOutlined } from '@ant-design/icons';
+import Icon, { UnorderedListOutlined, PlusOutlined, DeleteFilled, EllipsisOutlined, ApartmentOutlined } from '@ant-design/icons';
 import {NotePriority} from './NotePriority.jsx';
 import {NoteBacklinks} from './NoteBacklinks.jsx';
 import {NoteTags} from './NoteTags.jsx';
@@ -52,17 +52,6 @@ import contentUiCss from '!!raw-loader!tinymce/skins/ui/oxide/content.min.css';
 import contentNNCustomCss from '!!raw-loader!../css/tinymce.css';
 import { replace } from 'lodash';
 
-
-const TreeSvg = () => (
-    <svg width="24px" height="24px" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
-        <g>
-            <path fill="none" d="M0 0H24V24H0z"/>
-            <path d="M10 2c.552 0 1 .448 1 1v4c0 .552-.448 1-1 1H8v2h5V9c0-.552.448-1 1-1h6c.552 0 1 .448 1 1v4c0 .552-.448 1-1 1h-6c-.552 0-1-.448-1-1v-1H8v6h5v-1c0-.552.448-1 1-1h6c.552 0 1 .448 1 1v4c0 .552-.448 1-1 1h-6c-.552 0-1-.448-1-1v-1H7c-.552 0-1-.448-1-1V8H4c-.552 0-1-.448-1-1V3c0-.552.448-1 1-1h6zm9 16h-4v2h4v-2zm0-8h-4v2h4v-2zM9 4H5v2h4V4z"/>
-        </g>
-    </svg>
-);
-
-const TreeIcon = (props) => <Icon component={TreeSvg} {...props} />;
 
 class Note extends React.Component {
 
@@ -279,7 +268,7 @@ class Note extends React.Component {
             noteMenuItems.push({
                 key: 'open_tree',
                 label: "Open tree",
-                icon: <TreeIcon />,
+                icon: <ApartmentOutlined />,
             });
             noteMenuItems.push({
                 key: 'open_list',